Fix warning for unsigned int comparison on mac.

+++ pl2/tests/serializer.cc
@@ -38,14 +38,14 @@ TEST(FixedSerializer, SerializationTest) {
 
 TEST(VariableSerializer, EmptyRecordLengthTest) {
     Record *record = alloc_record();
-    ASSERT_EQ(sizeof(int) * SCHEMA_NUM_ATTRS, var_len_sizeof(record));
+    ASSERT_EQ(sizeof(int) * SCHEMA_NUM_ATTRS, (unsigned int) var_len_sizeof(record));
     free_record(record);
 }
 
 TEST(VariableSerializer, SparseRecordLengthTest) {
     Record *record = alloc_record();
     strcpy((char *) record->at(0), "hello");
-    ASSERT_EQ(sizeof(int) * SCHEMA_NUM_ATTRS + strlen("hello"), var_len_sizeof(record));
+    ASSERT_EQ(sizeof(int) * SCHEMA_NUM_ATTRS + strlen("hello"), (unsigned int) var_len_sizeof(record));
     free_record(record);
 }
 
@@ -57,7 +57,7 @@ TEST(VariableSerializer, FullRecordLengthTest) {
         sprintf((char *) record->at(i), "%9d", i);
     }
 
-    ASSERT_EQ(sizeof(int) * SCHEMA_NUM_ATTRS + 900, var_len_sizeof(record));
+    ASSERT_EQ(sizeof(int) * SCHEMA_NUM_ATTRS + 900, (unsigned int) var_len_sizeof(record));
     free_record(record);
 }
